// utils/Table/interface.ts
export type TableSize = "small" | "medium" | "large"

export type TableLayout = "auto" | "fixed"
export type TableOverFlow = "pagination" | "scroll"

export type TableAlign =
  | "left"
  | "center"
  | "right"
  | "start"
  | "end"
  | "flex-start"
  | "flex-end"

export type TableColorScheme =
  | string
  | "white"
  | "blackAlpha"
  | "gray"
  | "grayBlue"
  | "red"
  | "orange"
  | "yellow"
  | "green"
  | "blue"
  | "cyan"
  | "purple"
  | "techPink"
  | "techPurple"

export interface TableProps<D extends TableData, TValue>
  extends HTMLAttributes<HTMLDivElement>,
    TableContextProps,
    BoxProps {
  tableRef?: ForwardedRef<TableHandler<D>>
  colorScheme?: TableColorScheme
  columns?: ColumnDef<D, TValue>[]
  data?: D[]
  pinedHeader?: boolean
  tableLayout?: TableLayout
  bordered?: boolean
  hoverable?: boolean
  customCellPadding?: string
  defaultSort?: ColumnSort[]
  disableSortBy?: boolean
  enableColumnResizing?: boolean
  loading?: boolean
  emptyProps?: EmptyProps
  overFlow?: TableOverFlow
  pagination?: any
  multiRowSelection?: boolean
  enableRowSelection?: boolean
  enableSingleCellSelection?: boolean
  clickOutsideToResetRowSelect?: boolean
  serverSidePagination?: boolean
  // useAble when serverSidePagination is true
  total?: number
  checkAll?: boolean
  download?: boolean
  downloadRawData?: boolean
  refresh?: boolean
  filter?: boolean
  columnSizing?: ColumnSizingState
  rowSelection?: RowSelectionState
  columnVisibility?: VisibilityState
  onRefresh?: (table: Table<D>) => void
  onRowClick?: (row: Row<D>, index: number) => void
  onSortingChange?: (sort: SortingState) => void
  onGlobalFiltersChange?: (
    filters: FilterOptions[],
    operator: FilterOperator,
  ) => void
  onColumnFiltersChange?: OnChangeFn<ColumnFiltersState>
  onPaginationChange?: (
    paginationState: PaginationState,
    table: Table<D>,
  ) => void
  onRowSelectionChange?: (rowSelection?: RowSelectionState) => void
  onCellSelectionChange?: (cell?: Cell<D, TValue>) => void
  onColumnSizingChange?: (columnSizing?: ColumnSizingState) => void
}

export interface TableHandler<D extends TableData> {
  table: Table<D>
  selectPage: (pageIndex: number) => void
  selectRow: (rowSelection: RowSelectionState) => void
  setGlobalFilters: (filters: FilterOptions[], operator: FilterOperator) => void
  clearSelection: () => void
}

// eslint-disable-next-line @typescript-eslint/no-unused-vars
export interface RowSelectionProps<D = any> {
  checkAll?: boolean
}

export interface TableContextProps {
  borderedCell?: boolean
  striped?: boolean
  showHeader?: boolean
  showFooter?: boolean
  customCellPadding?: string
  size?: TableSize
  align?: TableAlign
  hoverable?: boolean
  enableColumnResizing?: boolean
}
// ...
export type FilterFn = FilterFnOption<any> & CustomFilterFn

export type FilterOptions = {
  id: string
  value: any
  filterFn?: FilterFn
}

export type FilterOptionsState = FilterOptions[]

export type FilterOperator = "and" | "or"